Feature/kaleb coberly/add extraction notebook (#3)
Adds exploration notebook, which outlines approach and starts in on image preprocessing and field extraction with OpenCV and PyMuPDF, but finds a hitch in setting up the template form image and fiducial corners.

Adds template form PDF and ideal-ish raw intake images, with an expected extraction doc.

Adds form version to metadata.